requirements.txt
# Flask y extensiones principales (ESENCIALES)
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-Migrate==4.0.5

# Dependencias de base de datos (ESENCIALES)
WTForms==3.1.1
PyMySQL==1.1.0
mysql-connector-python==8.1.0
SQLAlchemy==2.0.23

# Utilidades básicas (ESENCIALES)
python-dotenv==1.0.0
Werkzeug==2.3.7
itsdangerous==2.1.2
email-validator==2.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
blinker==1.7.0

# Generación de PDFs (ESENCIAL para reportes)
pdfkit==1.0.0
xhtml2pdf==0.2.11
Pillow==10.1.0

# Exportación a Excel (ESENCIAL para reportes)
openpyxl==3.1.2

# Migraciones de base de datos (ESENCIAL)
alembic==1.12.1

# Producción (ESENCIAL para deploy)
gunicorn==21.2.0

# ==========================================
# DEPENDENCIAS OPCIONALES (descomenta si las necesitas)
# ==========================================

# Análisis de datos avanzado (OPCIONAL)
# pandas==2.1.4
# numpy==1.25.2
# matplotlib==3.8.2

# Testing (OPCIONAL - solo para desarrollo)
# pytest==7.4.3
# pytest-cov==4.1.0
# coverage==7.3.2

# Seguridad adicional (OPCIONAL)
# cryptography==41.0.7
# bcrypt==4.1.2

# Utilidades adicionales (OPCIONAL)
# python-dateutil==2.8.2
# pytz==2023.3
# gevent==23.9.1
# python-magic==0.4.27
# structlog==23.2.0
# requests==2.31.0
# urllib3==2.1.0
# six==1.16.0
# xlrd==2.0.1 